requirements.txt
openai >= 1.97.1
g4f >= 0.5.7.6
fastapi>=0.116.1
uvicorn>=0.35.0
streamlit>=1.46.1
tavily-python>=0.7.10
python-dotenv>=1.1.1
langgraph>=0.6.1
ipython>=9.4.0
PyPDF2>=3.0.1
python-docx>=1.2.0
langchain-openai>=0.3.28
langchain>=0.3.26
av>=10.0.0
pydub>=0.25.1
wave
librosa>=0.10.0
soundfile>=0.12.0
mutagen>=1.46.0